+++ packages/next/CHANGELOG.md
@@ -21,7 +21,7 @@
 
 - 068f3c3: Introduced the `useLogout` hook to facilitate logging out a user
 
-  See https://faustjs.org/docs/next/reference/custom-hooks#uselogout for more details.
+  See https://faustjs.org/docs/next/reference/hooks/useLogout for more details.
 
 - Updated dependencies [068f3c3]
   - @faustjs/core@0.12.3